+++ js/avatar.js
@@ -1,0 +1,15 @@
+const FILE_TYPES = ['gif', 'jpg', 'jpeg', 'png'];
+
+const fileChooser = document.querySelector('.img-upload__start input[type=file]');
+const preview = document.querySelector('#img-upload__preview');
+
+fileChooser.addEventListener('change', () => {
+  const file = fileChooser.files[0];
+  const fileName = file.name.toLowerCase();
+
+  const matches = FILE_TYPES.some((elem) => fileName.endsWith(elem));
+
+  if (matches) {
+    preview.src = URL.createObjectURL(file);
+  }
+});
