+++ js/main.js
@@ -1,3 +1,4 @@
 import './bigPictures.js';
 import './form.js';
-
+import './avatar.js';
+import './filter.js';
